WC2-172: code review (1)
